refactoring integration tests
